pyproject.toml
[project]
name = "rider-receipt-merger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "msoffcrypto-tool>=5.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyside6>=6.9.3",
]
